Added support for Cortex-M0 SAMD21 series. (#18)
Builder:
None

Compiler:
Implemented complex64 and complex128 types.

Runtime:
Added runtime package for SAMD21.
Implemented Pin interface for SAMD21.
Implemented subset of atomic operations for targets with no native support.
Implemented common registers for Cortex-M in runtime/cortexm package.
Corrected stacking implementation for non-THUMB2 targets (ARMV6-M).
Generalized scheduler runtime for Cortex-M targets.
Created structs representing complex numbers.
Stubbed out reflect package.
Implemented strconv.FormatInt
Implemented strconv.Itoa

Other:

Known issues:
Closures sometimes create values that are not dominated.
Testing is still not optimal as the test output still produces too much garbage making tests less coherent.
SVD for ATSAMD21G18A produces unusable PORT registers.

+++ src/runtime/arm/cortexm/sam/atsamx51/sercom.go
@@ -121,7 +121,7 @@ func (s SERCOM) SetEnabled(enable bool) {
 }
 
 func (s SERCOM) Baud(hz uint) uint8 {
-	return uint8((SERCOM_REF_FREQUENCY / (2 * hz)) - 1)
+	return uint8((SERCOM_REF_FREQUENCY / (2 * uint32(hz))) - 1)
 }
 
 func (s SERCOM) BaudFP(hz uint) (uint16, uint8) {
